Reorganised mix_nodes module
