+++ explorer-api/src/country_statistics/geolocate.rs
@@ -1,9 +1,11 @@
+// Copyright 2022 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: Apache-2.0
+
+use crate::mix_nodes::location::{GeoLocation, Location};
+use crate::state::ExplorerApiStateContext;
 use log::{info, warn};
 use reqwest::Error as ReqwestError;
 use thiserror::Error;
-
-use crate::mix_nodes::{GeoLocation, Location};
-use crate::state::ExplorerApiStateContext;
 
 pub(crate) struct GeoLocateTask {
     state: ExplorerApiStateContext,
@@ -34,7 +36,7 @@ impl GeoLocateTask {
     }
 
     async fn locate_mix_nodes(&mut self) {
-        let mixnode_bonds = self.state.inner.mix_nodes.get().await.value;
+        let mixnode_bonds = self.state.inner.mix_nodes.get().await.all_mixnodes;
 
         for (i, cache_item) in mixnode_bonds.values().enumerate() {
             if self
